+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.2.0 (2019-06-17)
+- Added Boolean Algorithm to Quad Algortithmic Rhythm
+- Added Limited Ellipse to Roulette LFO (better model forthcoming)
+- Fixed bug in BPM LFO Square output
+
 ### 1.1.0 (2019-06-17)
 - Ability to sync to arbitrary clock ratios added to BPM LFO and BPM LFO2
 - Added Waveshape interpolation to BPM LFO2 and Lissajous LFO